(CDAP-16369) Move TrafficRelayServer and ProxySelector out from RuntimeMonitorServer
- This is for preparing the removal of RuntimeMonitorServer
